feat(markdown): split the lezer grammar layer into `@meowdown/markdown`
