+++ .release-please-manifest.json
@@ -1,4 +1,5 @@
 {
   "packages/core": "0.52.2",
+  "packages/markdown": "0.52.2",
   "packages/react": "0.52.2"
 }
