+++ package.json
@@ -8,7 +8,7 @@
     "dev": "pnpm --filter ./website run dev",
     "build": "pnpm -r run build",
     "preview": "pnpm run build && pnpm --filter ./website run preview",
-    "ci:publish:snapshot": "pkg-pr-new publish --pnpm ./packages/core ./packages/react",
+    "ci:publish:snapshot": "pkg-pr-new publish --pnpm ./packages/markdown ./packages/core ./packages/react",
     "lint": "pnpm run lint:eslint && pnpm run lint:oxfmt && pnpm run lint:knip",
     "lint:oxfmt": "oxfmt --check",
     "lint:eslint": "eslint --cache --max-warnings 0",
