+++ packages/core/package.json
@@ -26,7 +26,7 @@
     "@floating-ui/dom": "^1.8.0",
     "@lezer/common": "^1.5.2",
     "@lezer/highlight": "^1.2.3",
-    "@lezer/markdown": "^1.7.1",
+    "@meowdown/markdown": "workspace:^",
     "@ocavue/utils": "^1.7.0",
     "@prosekit/core": "^0.13.0-beta.6",
     "@prosekit/extensions": "^0.18.0-beta.16",
