+++ packages/markdown/src/wiki-embed.ts
@@ -7,7 +7,7 @@ import {
   CHAR_RIGHT_SQUARE_BRACKET,
   CHAR_SPACE,
   CHAR_TAB,
-} from '../unicode.ts'
+} from './unicode.ts'
 
 /**
  * Inline parser for Obsidian-style wiki embeds (`![[target]]`). The target is
