+++ packages/markdown/tsdown.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from 'tsdown'
+
+export default defineConfig({
+  entry: ['src/index.ts'],
+  platform: 'neutral',
+  sourcemap: 'hidden',
+  minify: true,
+  target: 'es2022',
+})
